+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGridItem.java
@@ -35,7 +35,7 @@ public class WirelessGridItem extends AbstractNetworkEnergyItem {
 
     public EnergyStorage createEnergyStorage(final ItemStack stack) {
         final EnergyStorage energyStorage = new EnergyStorageImpl(
-            Platform.INSTANCE.getConfig().getWirelessGrid().getEnergyCapacity()
+            Math.clamp(Platform.INSTANCE.getConfig().getWirelessGrid().getEnergyCapacity(), 1, Long.MAX_VALUE)
         );
         return RefinedStorageApi.INSTANCE.asItemEnergyStorage(energyStorage, stack);
     }