+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/ConfigImpl.java
@@ -24,6 +24,8 @@ import net.minecraft.resources.ResourceLocation;
 public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.common.Config {
     private ScreenSize screenSize = ScreenSize.STRETCH;
 
+    private boolean requireEnergy = true;
+
     private boolean smoothScrolling = true;
 
     private boolean tenthAnniversaryCape = false;
@@ -158,6 +160,11 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     public void setScreenSize(final ScreenSize screenSize) {
         this.screenSize = screenSize;
         AutoConfig.getConfigHolder(ConfigImpl.class).save();
+    }
+
+    @Override
+    public boolean isRequireEnergy() {
+        return requireEnergy;
     }
 
     @Override
@@ -560,8 +567,10 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class ControllerEntryImpl implements ControllerEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.CONTROLLER_CAPACITY;
 
+        @Override
         public long getEnergyCapacity() {
             return energyCapacity;
         }
@@ -714,6 +723,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class WirelessGridEntryImpl implements WirelessGridEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.WIRELESS_GRID_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.WIRELESS_GRID_OPEN;
@@ -760,6 +770,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class PortableGridEntryImpl implements PortableGridEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.PORTABLE_GRID_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.PORTABLE_GRID_OPEN;
@@ -861,6 +872,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class WirelessAutocraftingMonitorEntryImpl implements WirelessAutocraftingMonitorEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_OPEN;